backend/requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
python-dotenv==1.1.0
flask-jwt-extended==4.4.4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==1.4.52
psycopg2-binary==2.9.6